refactor: refactor ExtensionManager into CatalogCrawlerManager
