+++ core/crawler-core/build.gradle.kts
@@ -14,7 +14,6 @@
 
 plugins {
     `java-library`
-    `java-test-fixtures`
 }
 
 dependencies {
