--- core/crawler-core/src/test/java/org/eclipse/edc/catalog/cache/TestUpdateResponse.java
@@ -1,23 +1,0 @@
-/*
- *  Copyright (c) 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
- */
-
-package org.eclipse.edc.catalog.cache;
-
-import org.eclipse.edc.crawler.spi.model.UpdateResponse;
-
-public class TestUpdateResponse extends UpdateResponse {
-    public TestUpdateResponse(String source) {
-        super(source);
-    }
-}
