+++ core/federated-catalog-core/src/main/java/org/eclipse/edc/catalog/cache/FederatedCatalogCoreServicesExtension.java
@@ -14,20 +14,15 @@
 
 package org.eclipse.edc.catalog.cache;
 
-import org.eclipse.edc.catalog.spi.CatalogConstants;
+import org.eclipse.edc.catalog.crawler.CatalogCrawlerManager;
 import org.eclipse.edc.catalog.spi.CatalogCrawlerConfiguration;
 import org.eclipse.edc.catalog.spi.FederatedCatalogCache;
-import org.eclipse.edc.catalog.spi.model.CatalogUpdateResponse;
-import org.eclipse.edc.connector.controlplane.catalog.spi.Catalog;
 import org.eclipse.edc.crawler.spi.CrawlerActionRegistry;
 import org.eclipse.edc.crawler.spi.TargetNodeDirectory;
 import org.eclipse.edc.crawler.spi.TargetNodeFilter;
-import org.eclipse.edc.crawler.spi.model.ExecutionPlan;
-import org.eclipse.edc.crawler.spi.model.UpdateResponse;
 import org.eclipse.edc.runtime.metamodel.annotation.Configuration;
 import org.eclipse.edc.runtime.metamodel.annotation.Extension;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
-import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.spi.system.ServiceExtension;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
 import org.eclipse.edc.spi.system.health.HealthCheckResult;
@@ -43,8 +38,6 @@ public class FederatedCatalogCoreServicesExtension implements ServiceExtension {
     @Configuration
     private CatalogCrawlerConfiguration catalogCrawlerConfiguration;
 
-    @Inject(required = false)
-    private ExecutionPlan executionPlan;
     @Inject
     private FederatedCatalogCache store;
     @Inject
@@ -56,28 +49,21 @@ public class FederatedCatalogCoreServicesExtension implements ServiceExtension {
     @Inject(required = false)
     private HealthCheckService healthCheckService;
 
-    private Monitor monitor;
-    private ExecutionManager executionManager;
+    private CatalogCrawlerManager manager;
 
     @Override
     public void initialize(ServiceExtensionContext context) {
-        monitor = context.getMonitor();
-
         if (healthCheckService != null) {
             healthCheckService.addReadinessProvider(() -> HealthCheckResult.Builder.newInstance().component("Crawler Subsystem").build());
         }
 
         nodeFilter = ofNullable(nodeFilter).orElse(node -> !node.name().equals(context.getRuntimeId()));
 
-        executionManager = ExecutionManager.Builder.newInstance()
-                .monitor(context.getMonitor().withPrefix("ExecutionManager"))
+        manager = CatalogCrawlerManager.Builder.newInstance()
+                .monitor(context.getMonitor().withPrefix(CatalogCrawlerManager.class.getSimpleName()))
                 .configuration(catalogCrawlerConfiguration)
-                .preExecutionTask(() -> {
-                    store.deleteExpired();
-                    store.expireAll();
-                })
+                .store(store)
                 .nodeQueryAdapterRegistry(crawlerActionRegistry)
-                .onSuccess(this::persist)
                 .nodeDirectory(directory)
                 .nodeFilterFunction(nodeFilter)
                 .build();
@@ -85,26 +71,12 @@ public class FederatedCatalogCoreServicesExtension implements ServiceExtension {
 
     @Override
     public void start() {
-        executionManager.executePlan(executionPlan);
+        manager.start();
     }
 
     @Override
     public void shutdown() {
-        executionManager.shutdownPlan(executionPlan);
+        manager.stop();
     }
 
-    /**
-     * inserts a particular {@link Catalog} in the {@link FederatedCatalogCache}
-     *
-     * @param updateResponse The response that contains the catalog
-     */
-    private void persist(UpdateResponse updateResponse) {
-        if (updateResponse instanceof CatalogUpdateResponse catalogUpdateResponse) {
-            var catalog = catalogUpdateResponse.getCatalog();
-            catalog.getProperties().put(CatalogConstants.PROPERTY_ORIGINATOR, updateResponse.getSource());
-            store.save(catalog);
-        } else {
-            monitor.warning("Expected a response of type %s but got %s. Will discard".formatted(CatalogUpdateResponse.class, updateResponse.getClass()));
-        }
-    }
 }
