+++ core/federated-catalog-core/src/main/java/org/eclipse/edc/catalog/cache/FederatedCatalogDefaultServicesExtension.java
@@ -14,26 +14,20 @@
 
 package org.eclipse.edc.catalog.cache;
 
-import org.eclipse.edc.catalog.cache.crawler.CrawlerActionRegistryImpl;
 import org.eclipse.edc.catalog.cache.query.QueryServiceImpl;
-import org.eclipse.edc.catalog.crawler.RecurringExecutionPlan;
+import org.eclipse.edc.catalog.crawler.CrawlerActionRegistryImpl;
 import org.eclipse.edc.catalog.directory.InMemoryNodeDirectory;
-import org.eclipse.edc.catalog.spi.CatalogCrawlerConfiguration;
 import org.eclipse.edc.catalog.spi.FederatedCatalogCache;
 import org.eclipse.edc.catalog.spi.QueryService;
 import org.eclipse.edc.catalog.store.InMemoryFederatedCatalogCache;
 import org.eclipse.edc.crawler.spi.CrawlerActionRegistry;
 import org.eclipse.edc.crawler.spi.TargetNodeDirectory;
-import org.eclipse.edc.crawler.spi.model.ExecutionPlan;
 import org.eclipse.edc.query.CriterionOperatorRegistryImpl;
-import org.eclipse.edc.runtime.metamodel.annotation.Configuration;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
 import org.eclipse.edc.runtime.metamodel.annotation.Provider;
 import org.eclipse.edc.spi.system.ServiceExtension;
-import org.eclipse.edc.spi.system.ServiceExtensionContext;
 import org.eclipse.edc.util.concurrency.LockManager;
 
-import java.time.Duration;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 /**
@@ -43,9 +37,6 @@ import java.util.concurrent.locks.ReentrantReadWriteLock;
 public class FederatedCatalogDefaultServicesExtension implements ServiceExtension {
 
     public static final String NAME = "Federated Catalog Default Services";
-
-    @Configuration
-    private CatalogCrawlerConfiguration catalogCrawlerConfiguration;
 
     @Inject
     private FederatedCatalogCache store;
@@ -75,13 +66,4 @@ public class FederatedCatalogDefaultServicesExtension implements ServiceExtensio
         return new CrawlerActionRegistryImpl();
     }
 
-    @Provider(isDefault = true)
-    public ExecutionPlan createRecurringExecutionPlan(ServiceExtensionContext context) {
-        var monitor = context.getMonitor();
-        catalogCrawlerConfiguration.checkPeriodSeconds().ifPresent(monitor::warning);
-        return new RecurringExecutionPlan(
-                Duration.ofSeconds(catalogCrawlerConfiguration.periodSeconds()),
-                Duration.ofSeconds(catalogCrawlerConfiguration.delaySeconds()),
-                monitor);
-    }
 }
