--- spi/crawler-spi/src/main/java/org/eclipse/edc/crawler/spi/package-info.java
@@ -1,18 +1,0 @@
-/*
- *  Copyright (c) 2020 - 2022 Microsoft Corporation
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Microsoft Corporation - initial API and implementation
- *
- */
-
-@Spi("Crawler services")
-package org.eclipse.edc.crawler.spi;
-
-import org.eclipse.edc.runtime.metamodel.annotation.Spi;
